move controllers and dto to different folders
