+++ Quiz/ComplexityWeb/QuizService/DTO/TopicInfoDTO.cs
@@ -1,7 +1,7 @@
 using System;
 using Newtonsoft.Json;
 
-namespace ComplexityWebApi.DTO
+namespace QuizWebApp.QuizService.DTO
 {
     public class TopicInfoDTO
     {
