+++ Quiz/ComplexityWeb/TaskService/DTO/TemplateGeneratorForRenderDTO.cs
@@ -1,6 +1,6 @@
 using Newtonsoft.Json;
 
-namespace QuizWebApp.DTO
+namespace QuizWebApp.TaskService.DTO
 {
     public class TemplateGeneratorForRenderDTO
     {
